% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dsts.R
\name{check_dsts}
\alias{check_dsts}
\title{Check distributions input}
\usage{
check_dsts(distributions)
}
\arguments{
\item{distributions}{List of objects to check. Should be flattened if
coming from \code{...}.}
}
\value{
Invisible; ran for the potential error.
}
\description{
Checks whether a collection of objects are either distributions,
NA, or NULL. If one object does not satisy these requirements,
an error is thrown.
}
