% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution2.R
\name{distribution2}
\alias{distribution2}
\title{Make a distribution from scratch}
\usage{
distribution2(
  ...,
  .vtype = c("continuous", "discrete", "mixed"),
  parameters = params(),
  parenv = rlang::caller_env()
)
}
\arguments{
\item{parameters}{Vector of unquoted variable names acting as the
distribution's parameters, if relevant.}

\item{variable}{Is this variable continuous, discrete, or mixed?}

\item{env}{Either the package name where Environment containing objects
needed for evaluating the distribution.}
}
\value{
A distribution object with nothing in it.
}
\description{
Initiate a distribution object by specifying the representation space.
}
\details{
This function sets up a distribution object with high-level
information about the distribution. You can add / modify information about
the distribution downstream using the \verb{set_*()} family of functions.

This becomes the global environment for the function, so that when an object
is needed to be evaluated, it ultimately looks here.
}
