% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_bern.R
\name{dst_bern}
\alias{dst_bern}
\title{Bernoulli Distribution}
\usage{
dst_bern(prob)
}
\arguments{
\item{prob}{Probability of success; single numeric between 0 and 1.}
}
\value{
A Bernoulli distribution.
}
\description{
Makes a Bernoulli distribution, representing the outcome of a
single trial with a given success probability.
}
\examples{
dst_bern(0.3)
}
