% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_beta.R
\name{dst_beta}
\alias{dst_beta}
\title{Beta Distribution}
\usage{
dst_beta(shape1, shape2)
}
\arguments{
\item{shape1, shape2}{Shape parameters of the distribution;
single positive numerics.}
}
\value{
A Beta distribution.
}
\description{
Makes a Beta distribution.
}
\examples{
dst_beta(2, 3)
}
