% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_binom.R
\name{dst_binom}
\alias{dst_binom}
\title{Binomial Distribution}
\usage{
dst_binom(size, prob)
}
\arguments{
\item{size}{Number of trials.}

\item{prob}{Success probability of each trial.}
}
\value{
A binomial distribution.
}
\description{
Makes a Binomial distribution, representing the number of successes
in a fixed number of independent trials.
}
\examples{
dst_binom(10, 0.6)
}
