% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_cauchy.R
\name{dst_cauchy}
\alias{dst_cauchy}
\title{Cauchy Distribution}
\usage{
dst_cauchy(location, scale)
}
\arguments{
\item{location}{Location parameter; single numeric.}

\item{scale}{Scale parameter; single positive numeric.}
}
\value{
A Cauchy distribution.
}
\description{
Makes a Cauchy distribution.
}
\examples{
d <- dst_cauchy(0, 1)

# Moments do not exist for the Cauchy distribution.
mean(d)
variance(d)
}
