% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_chisq.R
\name{dst_chisq}
\alias{dst_chisq}
\title{Chi-Squared Distribution}
\usage{
dst_chisq(df)
}
\arguments{
\item{df}{degrees of freedom parameter; single positive numeric.}
}
\value{
A Chi-Squared distribution
}
\description{
Makes a Chi-Squared distribution.
}
\examples{
dst_chisq(3)
}
