% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_degenerate.R
\name{dst_degenerate}
\alias{dst_degenerate}
\title{Degenerate Distribution}
\usage{
dst_degenerate(location)
}
\arguments{
\item{location}{Single outcome of the distribution.}
}
\value{
A degenerate distribution
}
\description{
A degenerate distribution assigns a 100\% probability to one outcome.
}
\examples{
d <- dst_degenerate(5)
realise(d)
variance(d)
}
