% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_f.R
\name{dst_f}
\alias{dst_f}
\title{F Distribution}
\usage{
dst_f(df1, df2)
}
\arguments{
\item{df1, df2}{Degrees of freedom of the numerator and denominator,
both positive.}
}
\value{
An F distribution.
}
\description{
Makes an F distribution.
}
\examples{
dst_f(2, 3)
}
