% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_finite.R
\name{dst_finite}
\alias{dst_finite}
\title{Finite Distribution}
\usage{
dst_finite(outcomes, probs)
}
\arguments{
\item{outcomes}{Numeric vector representing the potential
outcomes of the distribution.}

\item{probs}{Numeric vector of probabilities corresponding to the outcomes
in \code{outcomes}. Must not be negative and must sum to 1.}
}
\value{
A distribution with finite outcomes.
}
\description{
Makes a finite distribution, which is a distribution with a finite number
of possible outcomes.
}
\examples{
dst_finite(2:5, probs = 1:4 / 10)
}
\seealso{
\code{\link[=dst_empirical]{dst_empirical()}}
}
