% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_gamma.R
\name{dst_gamma}
\alias{dst_gamma}
\title{Gamma Distribution}
\usage{
dst_gamma(shape, rate)
}
\arguments{
\item{shape}{Shape parameter; positive.}

\item{rate}{Rate parameter; positive.}
}
\value{
A Gamma distribution.
}
\description{
Makes a Gamma distribution.
}
\examples{
dst_gamma(2, 1)
}
