% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_geom.R
\name{dst_geom}
\alias{dst_geom}
\title{Geometric Distribution}
\usage{
dst_geom(prob)
}
\arguments{
\item{prob}{Probability of success in each trial; 0 < \code{p} <= 1.}
}
\description{
Makes a Geometric distribution, corresponding to the number of failures
in a sequence of independent trials before observing a success.
}
\examples{
d <- dst_geom(0.4)

# This version of the Geometric distribution does not count the success.
range(d)
}
