% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_gpd.R
\name{dst_gpd}
\alias{dst_gpd}
\title{Generalized Pareto Distribution}
\usage{
dst_gpd(scale, shape)
}
\arguments{
\item{scale}{Scale parameter; positive numeric.}

\item{shape}{Shape parameter; numeric.}
}
\value{
Object of class "dst" of a GPD.
}
\description{
Makes a distribution belonging to the family of
generalized Pareto distributions (GPD).
}
\examples{
dst_gpd(1, 1)
}
