% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_gumbel.R
\name{dst_gumbel}
\alias{dst_gumbel}
\title{Gumbel Distribution}
\usage{
dst_gumbel(location, scale)
}
\arguments{
\item{location}{Location parameter; single numeric.}

\item{scale}{Scale parameter; single positive numeric.}
}
\value{
A Gumbel distribution.
}
\description{
Makes a Gumbel distribution, which is a special case of the
Generalised Extreme Value (GEV) distribution when the shape parameter is 0.
}
\examples{
dst_gumbel(0, 1)
}
