% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_lp3.R
\name{dst_lp3}
\alias{dst_lp3}
\title{Log Pearson Type III distribution}
\usage{
dst_lp3(meanlog, sdlog, skew)
}
\arguments{
\item{meanlog, sdlog, skew}{Parameters.}
}
\value{
A Log Pearson Type III distribution.
}
\description{
Log Pearson Type III distribution
}
