% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_nbinom.R
\name{dst_nbinom}
\alias{dst_nbinom}
\title{Negative binomial Distribution}
\usage{
dst_nbinom(size, prob)
}
\arguments{
\item{size}{Number of successful trials; positive.}

\item{prob}{Probability of a successful trial; between 0 and 1.}
}
\value{
A Negative Binomial distribution.
}
\description{
Makes a Negative Binomial distribution, corresponding to the number
of failures in a sequence of independent trials until a given number
of successes are observed.
}
\examples{
d <- dst_nbinom(10, 0.5)

# This version of the Negative Binomial distribution does not count
# the successes.
range(d)
}
