% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_pois.R
\name{dst_pois}
\alias{dst_pois}
\title{Poisson Distribution}
\usage{
dst_pois(lambda)
}
\arguments{
\item{lambda}{Mean of the Poisson distribution; positive.}
}
\value{
A Poisson distribution.
}
\description{
Makes a Poisson distribution.
}
\examples{
dst_pois(1)
}
