% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_weibull.R
\name{dst_weibull}
\alias{dst_weibull}
\title{Weibull Distribution}
\usage{
dst_weibull(shape, scale)
}
\arguments{
\item{shape}{Shape parameter; single positive numeric.}

\item{scale}{Scale parameter; single positive numeric.}
}
\value{
A Weibull distribution.
}
\description{
Makes a Weibull distribution.
}
\examples{
dst_weibull(1, 1)
}
