% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_from_network-quantile.R
\name{encapsulate_p}
\alias{encapsulate_p}
\title{Find a range of possible outcomes}
\usage{
encapsulate_p(distribution, p, direction)
}
\arguments{
\item{distribution}{A distribution having access to a cdf.}

\item{p}{Vector of values between 0 and 1 (inclusive).}

\item{direction}{One of \code{"left"} for calculating left-inverse, or
\code{"right"} for calculating right-inverse.}
}
\value{
A range of values containing the solutions to the left
inverse of the CDF at \code{p}.
}
\description{
In order to run the directional inverse algorithm, we need to know
approximately where the solution lies. This function
finds a range of possible outcomes where the cdf evaluates to values
(probabilities) contain the vector \code{p}, and therefore should come
before the inversion algorithm begins.
}
\note{
If 0 or 1 are included in the vector \code{p}, one of the endpoints might
be infinite.
}
\examples{
d <- distribution(cdf = \(x) pmin(pmax(x, 0), 1))
distionary:::encapsulate_p(d, p = 1:9 / 10, direction = "left")

}
