% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe_cdf.R
\name{enframe_cdf}
\alias{enframe_cdf}
\alias{enframe_bi_cdf}
\alias{enframe_multi_cdf}
\title{Enframe a cumulative distribution function (CDF)}
\usage{
enframe_cdf(..., at, arg_name = ".arg", fn_prefix = "cdf", sep = "_")

enframe_bi_cdf(..., x, y, arg_name = NULL, fn_prefix = "cdf", sep = "_")

enframe_multi_cdf(..., .l, arg_name = NULL, fn_prefix = "cdf", sep = "_")
}
\arguments{
\item{...}{Distributions, or possibly a nested list of distributions,
inheriting the class \code{"dst"}, \code{"bidst"}, or \code{"multidst"}.
Must be of the same dimension.}

\item{arg_name}{Name(s) of the column(s) containing
the function arguments, one per dimension of the distribution. See details.}

\item{fn_prefix}{Name of the prefix to appear in the function
evaluation column(s).}

\item{sep}{Deprecated; no longer used.}
}
\value{
A data frame, or a tibble if the tibble package is accessible, with
one column per argument (one in the case of \code{at}; two in the case
of \code{x}, \code{y}), and one column per distribution input into \code{...}.
}
\description{
Evaluates the CDF of the specified distributions, placing the inputs
and outputs in a data frame.
}
\details{
Column names can be specified directly through
\code{arg_name} (for the input columns) and \code{fn_prefix} (for the distributions /
output columns), one name per column. If only one name is supplied, this
name is included as a prefix to the
automated naming convention, which arrives at names using the following
preferential order:
\enumerate{
\item Given names, where applicable (as in \code{...} and \code{.l}).
\item Variable names input as arguments.
\item \code{.x} and \code{.y}, in the case of bivariate enframing, or \code{...1}, \code{...2},
etc. in the case of multivariate enframing.
}
}
