% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe_general.R
\name{enframe_general}
\alias{enframe_general}
\title{Enframe a distributional representation's evaluation}
\usage{
enframe_general(..., .l, fn_prefix, eval_fn, fn_args = list())
}
\arguments{
\item{...}{Distributions, or possibly a nested list of distributions,
inheriting the class \code{"dst"}, \code{"bidst"}, or \code{"multidst"}.
Must be of the same dimension.}

\item{.l}{A list of vectors, one for each dimension of the distribution,
to evaluate the distributional representation at. Optionally, a single
vector of the same length, for evaluating the representation at a single
point.}

\item{fn_prefix}{Name of the prefix to appear in the function
evaluation column(s).}

\item{eval_fn}{Character name of the eval function to enframe;
e.g., \code{"eval_multi_cdf"}.}

\item{fn_args}{A named list of arguments to pass to the \code{eval_fn} function,
besides the distribution and \code{.l} argument (the \code{strict} argument
being the most common, and perhaps the only use case).}
}
\value{
A data frame or tibble of the input argument(s) (\code{.l}), with the
evaluated distributional representation for each distribution in
\code{...} in its own column.
}
\description{
These are the workhorses for the \code{enframe_} family of functions, evaluating
a specified distributional representation for univariate, bivariate, and
multivariate distributions, for multiple distributions,
and places the results in a data frame or tibble.
}
\details{
If only one distribution is specified in \code{...}, then the evaluation
column will be named that of \code{fn_prefix}.

If more than one distribution
is specified in \code{...}, the evaluation columns will be named by the
prefix \code{fn_prefix} followed by the distribution names.

Distributions are named first by their argument names, if given, or if not,
the input text. Names are then made unique using \code{vctrs::vec_as_names()}
with the "unique" names repair. "Unique" is chosen instead of "universal"
because names are anticipated to be syntactic with the \code{eval_fn} prefix;
"minimal" is not sufficient because it may result in columns having the
same names.
}
