% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R
\name{enframed_names_output}
\alias{enframed_names_output}
\alias{enframed_names_input}
\title{Column names for enframe functions}
\usage{
enframed_names_output(..., fn_prefix, sep)

enframed_names_input(input_names, arg_name)
}
\arguments{
\item{...}{Arguments contained in the ellipsis of an enframe_ function.}

\item{input_names}{A character vector of  input vectors,
on which the distributional
representation is to eventually be evaluated at.}
}
\description{
Derive names for the columns in the enframe_ family of functions;
"input" being the function argument being evaluated, and
"output" being the evaluated function.
}
