% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_binding.R
\name{env_bind_representations}
\alias{env_bind_representations}
\title{Bind to a representation environment}
\usage{
env_bind_representations(.env, ...)
}
\description{
Evaluates everything in ... within the specified environment if possible
and if non-NULL. Expressions in ... are allowed to refer to each other.
}
\details{
This function is like building an R package, but tries to allow for
the specification of constants. This may not be a good idea, for the
same reason that
"Any R code outside of a function is suspicious and should be
carefully reviewed" (https://r-pkgs.org/code.html#sec-code-when-executed).
An alternative may be to only use functions, e.g., \code{g <- function() 9.81}.

Evaluation occurs in order of appearance in \code{...}. To allow
earlier quantities to refer to later ones, the evaluate-and-bind
procedure is looped until there is nothing new to bind.
}
