% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_eval_and_bind.R
\name{env_eval_and_bind}
\alias{env_eval_and_bind}
\title{Evaluate and bind in an environment}
\usage{
env_eval_and_bind(.env, ...)
}
\arguments{
\item{.env}{An environment.}

\item{...}{Named objects to evaluate and bind in \code{.env}.}
}
\value{
The updated version of the environment \code{.env}, invisibly.
}
\description{
Like \emph{rlang}'s \code{env_bind} family of functions, this function also
takes named values and binds them to the specified environment, but also
evaluates the expressions in the environment. This has implications when
defining functions, because the encapsulating environment becomes the
input environment, not the caller environment.
}
\examples{
library(rlang)
e1 <- env()
e2 <- env()
env_bind(e1, f = \(x) x^2)
env_eval_and_bind(e2, f = \(x) x^2)
f_env(e1$f)
f_env(e2$f)
}
