% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bi_cdf.R
\name{eval_bi_cdf}
\alias{eval_bi_cdf}
\title{Convenience Functions for Bivariate Distributions}
\usage{
eval_bi_cdf(distribution, x, y, ...)
}
\arguments{
\item{distribution}{A distribution, inheriting the class \code{"dst"}.}

\item{x, y}{Numeric vectors specifying where to evaluate the bivariate
representation at.}
}
\description{
Evaluate bivariate distributions using \code{x} and \code{y} arguments, instead
of providing a list in the \code{at} argument of \verb{eval_<representation>()}.
}
