% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cdf.R, R/eval_bi_cdf.R, R/eval_multi_cdf.R
\name{eval_cdf}
\alias{eval_cdf}
\alias{eval_bi_cdf}
\alias{eval_multi_cdf}
\title{Cumulative Distribution Function}
\usage{
eval_cdf(distribution, at)

eval_bi_cdf(distribution, x, y)

eval_multi_cdf(distribution, .l)
}
\arguments{
\item{distribution}{A distribution, inheriting the class
\code{"dst"}, \code{"bidst"}, or \code{"multidst"}.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}

\item{x, y}{Vectors of values to evaluate the bivariate distributional
representation at.}

\item{.l}{A list of vectors, one for each dimension of the distribution,
to evaluate the distributional representation at. Optionally, a single
vector of the same length, for evaluating the representation at a single
point.}
}
\value{
The evaluated cdf in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Evaluate a cumulative distribution function (cdf).
}
\examples{
d1 <- dst_unif(0, 4)
d2 <- dst_pois(1.1)
eval_cdf(d1, at = 0:4)
enframe_cdf(d1, at = 0:4)
enframe_cdf(d1, d2, at = 0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_chf}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
