% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-cdf.R
\name{eval_cdf_clean}
\alias{eval_cdf_clean}
\title{Evaluate a representation using checked inputs}
\usage{
eval_cdf_clean(distribution, at)
}
\arguments{
\item{distribution}{A distribution, inheriting the class \code{"dst"}.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}
}
\description{
This is an internal generic that dispatches the appropriate representation
and anticipates cleaned inputs made by the user-facing functions without
the \verb{_clean} suffix.
}
\details{
Inputs are expected to be processed:
\itemize{
\item Dimension of the distribution must match the length of \code{at}.
\item If the distribution has dimension 1, \code{at} is an atomic vector;
if dimension >1, \code{at} is a list of vectors.
\item Vectors in \code{at} must be recycled to the same length.
}
}
