% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bi_density.R, R/eval_density.R,
%   R/eval_multi_density.R
\name{eval_bi_density}
\alias{eval_bi_density}
\alias{eval_density}
\alias{eval_multi_density}
\title{Probability Density Function}
\usage{
eval_bi_density(distribution, x, y, strict = TRUE, ...)

eval_density(distribution, at, strict = TRUE, ...)

eval_multi_density(distribution, .l, strict = TRUE, ...)
}
\arguments{
\item{distribution}{A distribution, inheriting the class
\code{"dst"}, \code{"bidst"}, or \code{"multidst"}.}

\item{x, y}{Vectors of values to evaluate the bivariate distributional
representation at.}

\item{strict}{Only evaluate when the density exists? \code{TRUE} if so;
if \code{FALSE}, evaluates the derivative of the cdf.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}

\item{.l}{A list of vectors, one for each dimension of the distribution,
to evaluate the distributional representation at. Optionally, a single
vector of the same length, for evaluating the representation at a single
point.}
}
\description{
Evaluate a distribution's probability density function (PDF).
}
\seealso{
\code{\link[=eval_cdf]{eval_cdf()}}
}
