% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_multi_cdf.R
\name{eval_multi_cdf.dst}
\alias{eval_multi_cdf.dst}
\title{Evaluate a univariate distribution in a multivariate framework}
\usage{
\method{eval_multi_cdf}{dst}(distribution, .l)
}
\description{
This method allows a user to use a multivariate evaluation function for
a univariate distribution.
}
