% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile_from_cdf.R
\name{eval_quantile_from_cdf}
\alias{eval_quantile_from_cdf}
\title{Evaluate Quantiles using a CDF}
\usage{
eval_quantile_from_cdf(distribution, at, tol, maxiter)
}
\arguments{
\item{distribution}{A distribution having access to a cdf.}

\item{at}{A vector of values for which to evaluate the quantile function.}

\item{tol, maxiter}{Tolerance and maximum number of iterations}
}
\description{
Intended for internal use only.
}
