% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_representation.R
\name{eval_representation}
\alias{eval_representation}
\title{Evaluate a distributional representation}
\usage{
eval_representation(distribution, entry, ...)
}
\arguments{
\item{distribution}{Distribution object.}

\item{entry}{Name of the representation / property, such as "cdf" or
"mean".}

\item{...}{If the representation is a function, arguments to the
function go here. Need not be named.}
}
\value{
The distribution's representation, evaluated. If cannot be
evaluated, returns \code{NULL}.
}
\description{
Evaluate a distributional representation. The distribution itself
is first searched for the representation / property, and if can't
be found, will attempt to calculate the representation / property
from other representations.
}
\examples{
d <- distribution(cdf = \(x) (x > 0) * pmin(x^2, 1), g = 9.81)
eval_representation(d, "g")
eval_representation(d, "quantile", 1:9 / 10)
eval_representation(d, "mean")
eval_representation(d, "realise", 10)
eval_representation(d, "foofy")
eval_representation(d, "foofy", 1:10)
}
