% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation-cdf.R, R/eval_pmf-generics.R,
%   R/representation-density.R
\name{eval_cdf}
\alias{eval_cdf}
\alias{eval_pmf}
\alias{eval_density}
\title{Evaluate a representation of a distribution}
\usage{
eval_cdf(distribution, at)

eval_pmf(distribution, ...)

eval_density(distribution, at, strict = TRUE, ...)
}
\arguments{
\item{distribution}{A distribution, inheriting the class \code{"dst"}.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}
}
\value{
The evaluated representation as a vector of length matching the
length of \code{at}.
}
\description{
These eval functions evaluate different representations of
univariate (1-variable) distributions.
}
\details{
Representations include:
\itemize{
\item \code{eval_cdf()} evaluates the cumulative distribution function.
\item \code{eval_survival()} evaluates the survival function (one minus the cdf).
\item \code{eval_density()} evaluates the probability density function.
\item \code{eval_pmf()} evaluates the probability mass function.
\item \code{eval_hazard()} evaluates the hazard function, for continous distributions.
\item \code{eval_chf()} evaluates the cumulative hazard function.
\item \code{eval_quantile()} evaluates the quantile function.
\item \code{eval_odds()} evaluates the odds function (\code{pmf / (1 - pmf)}).
\item \code{eval_return()} evaluates quantiles having a specified return period
(evaluating the return function at \code{rp} is the same as evaluating the
quantile function at \code{1 - 1 / rp}).
}
}
\examples{
eval_cdf(dst_norm(0, 1), at = -1:2)
eval_hazard(dst_exp(3.5), at = 0:4)
eval_chf(dst_weibull(1.3, 0.4), at = 0:4)
}
