% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_bind_within.R
\name{eval_within_if_possible}
\alias{eval_within_if_possible}
\title{Try to evaluate an expression within an environment}
\usage{
eval_within_if_possible(expr, env)
}
\description{
If an expression evaluates within an environment and is not null,
this function returns that value. Otherwise, the original expression
is returned. Because the evaluation occurs within the environment,
the environment becomes the enclosing environment when an expression
evaluates to a function.
}
\note{
This function could have been \code{eval_within_if_function()}, so that
the successful evaluation would also have to be a function in order to
be returned, but that would prevent functions from referencing
user-defined constants along the lines of Base R's \code{pi} and \code{letters}.
At the same time, expressions that rely on parameters will not evaluate
because a data mask is not used.
}
