% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evi.R
\name{evi}
\alias{evi}
\title{Extreme Value Index}
\usage{
evi(distribution)
}
\arguments{
\item{distribution}{Distribution to obtain EVI from.}
}
\value{
A single numeric.
}
\description{
Extreme Value Index
}
\examples{
evi(dst_gpd(0, 1, 2))
evi(dst_gpd(0, 1, -1))
evi(dst_unif(0, 10))
evi(dst_norm(5, 5))
evi(dst_empirical(1:10))
}
