% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdqr_gev.R
\name{gev_t_function}
\alias{gev_t_function}
\title{'t()' function for calculating GEV quantities}
\usage{
gev_t_function(x, location, scale, shape)
}
\arguments{
\item{x}{Argument of the function; vectorized.}

\item{location}{Location parameter; numeric.}

\item{scale}{Scale parameter; positive numeric.}

\item{shape}{Shape parameter; numeric. Also the extreme value index,
so that \code{shape > 0} is heavy tailed, and \code{shape < 0} is short-tailed.}
}
\value{
A vector of the t function evaluated.
}
\description{
't()' function for calculating GEV quantities
}
\note{
The shape parameter is not vectorized. This function is only
intended to be used when location, scale, and shape are scalars.
}
\examples{
gev_t_function(1:10, 0, 1, 1)

}
\seealso{
See the Wikipedia entry for the GEV distribution,
https://en.wikipedia.org/wiki/Generalized_extreme_value_distribution
}
