% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdq_gpd.R
\name{pgpd}
\alias{pgpd}
\alias{qgpd}
\alias{dgpd}
\title{Representations of the Generalized Pareto Distribution}
\usage{
pgpd(q, scale, shape, lower.tail = TRUE)

qgpd(p, scale, shape)

dgpd(x, scale, shape)
}
\arguments{
\item{scale}{Vector of scale parameters; positive numeric.}

\item{shape}{Vector of shape parameters; positive numeric.}

\item{lower.tail}{Single logical. If \code{TRUE}, cdf (default);
if \code{FALSE}, survival function.}

\item{p}{Vector of probabilities.}

\item{x, q}{Vector of quantiles.}
}
\value{
Vector of evaluated GPD distribution, with length
equal to the recycled lengths of \code{q}/\code{x}/\code{p}, \code{scale}, and \code{shape}.
}
\description{
Representations of the Generalized Pareto Distribution
}
\examples{
pgpd(1:10, 1, 1)
dgpd(1:10, 2, 0)
qgpd(1:9 / 10, 10, -2)
}
