% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_distribution.R
\name{is_distribution}
\alias{is_distribution}
\alias{is_dst}
\alias{is_bidst}
\alias{is_multidst}
\alias{is.distribution}
\title{Distribution Objects}
\usage{
is_distribution(object)

is_dst(object)

is_bidst(object)

is_multidst(object)

is.distribution(object)
}
\arguments{
\item{object}{Object to be tested}
}
\description{
Check if an object is a distribution.
}
