% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-quantile.R
\name{left_inverse}
\alias{left_inverse}
\title{Algorithm to Compute Left Inverse}
\usage{
left_inverse(distribution, p, low, high, tol, maxiter)
}
\arguments{
\item{low, high}{Single numeric values forming a range
within which to search for the solution.}

\item{tol}{Tolerance for the solution. The actual
solution will be within plus or minus half this value.}

\item{maxiter}{Number of iterations to attempt before
quitting, if the tolerance has not been reached. by then.}

\item{f}{A function to compute the left inverse of.}

\item{y}{Single value for which to calculate the left inverse.
Not vectorized.}
}
\description{
Calculates the smallest value for which a function
\code{f} evaluates to be greater than or equal to
\code{y} -- that is, the left inverse of \code{f}
at \code{y}. Intended for internal use only.
}
\details{
This algorithm works by progressively
cutting the specified range in half, so that the width
of the range after k iterations is 1/2^k times the
original width.
}
