% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdqr_lp3.R
\name{dlp3}
\alias{dlp3}
\alias{qlp3}
\alias{qplp3}
\alias{rlp3}
\title{LP3 distribution quantities}
\usage{
dlp3(x, meanlog, sdlog, skew)

qlp3(p, meanlog, sdlog, skew)

qplp3(q, meanlog, sdlog, skew)

rlp3(n, meanlog, sdlog, skew)
}
\arguments{
\item{x, q}{Vector of quantiles/magnitudes.}

\item{meanlog, sdlog, skew}{Parameters.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations to draw.}
}
\description{
p/d/q/r functions for the Log Pearson Type III distribution.
}
