% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{median.dst}
\alias{median.dst}
\title{Median of a Distribution}
\usage{
\method{median}{dst}(x, ...)
}
\arguments{
\item{x}{Distribution to calculate median from.}

\item{...}{Not used.}
}
\value{
Median of a distribution; single numeric.
}
\description{
Finds the median of a distribution.
}
\details{
Median is calculated as the 0.5-quantile when not found in the
distribution. So, when the median is
non-unique, the smallest of the possibilities is taken.
}
