% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cdf.R
\name{eval_bicdf}
\alias{eval_bicdf}
\alias{eval_multicdf}
\title{Multivariate Cumulative Distribution Function}
\usage{
eval_bicdf(distribution, x, y, ...)

eval_multicdf(distribution, .l, ...)
}
\arguments{
\item{x, y}{Vectors of values to evaluate the bivariate cdf at.}

\item{...}{Other arguments, currently not used, for extension.}

\item{.l}{List of vectors of values to evaluate the multivariate cdf at.}
}
\value{
The evaluated cdf in vector form.
}
\description{
Access a multivariate distribution's cumulative distribution function (cdf).
Bivariate function is provided for convenience.
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_chf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
