% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-quantile.R
\name{narrow_by_discretes}
\alias{narrow_by_discretes}
\title{Narrow a range of possible values by a discrete}
\usage{
narrow_by_discretes(distribution, low, high, discrete)
}
\description{
Narrow a range of possible values by a discrete
}
