% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enframe.R
\name{new_enframe_univariate}
\alias{new_enframe_univariate}
\alias{new_enframe_bivariate}
\alias{new_enframe_multivariate}
\title{Make a new \code{enframe_} function}
\usage{
new_enframe_univariate(repres)

new_enframe_bivariate(repres)

new_enframe_multivariate(repres)
}
\arguments{
\item{repres}{String; name of the representation.}
}
\value{
A function that evaluates the specified representation of inputted
distributions, and places the results in a data frame.
\itemize{
\item The \verb{_univariate} flavour provides the \code{at} argument to place a single
vector to evaluate the representation, as in \code{enframe_cdf()}.
\item The \verb{_bivariate} flavour provides the \code{x} and \code{y} arguments to hold
vectors to evaluate the representation, as in \code{enframe_bicdf()}.
\item The \verb{_multivariate} flavour provides the \code{.l} argument to hold a list of
vectors to evaluate the representation, as in \code{enframe_multicdf()}.
}
}
\description{
For a given distributional representation, these functions create
an \code{enframe_} function for that representation.
}
