% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution2.R
\name{params}
\alias{params}
\title{Parameters are a named list, the names being the parameters,
and the contents of each parameter entry being expressions defining
the parameter space. COMPLICATION: space defined with more than one
parameter together.}
\usage{
params(...)
}
\description{
Parameters are a named list, the names being the parameters,
and the contents of each parameter entry being expressions defining
the parameter space. COMPLICATION: space defined with more than one
parameter together.
}
