% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_parametric.R
\name{pdq_functions_exist}
\alias{pdq_functions_exist}
\title{Are there p/d/q functions available?}
\usage{
pdq_functions_exist(name)
}
\arguments{
\item{name}{Name of the distribution; suffix to p/d/q.}
}
\value{
\code{TRUE} if p/d/q functions are all available for this distribution;
\code{FALSE} otherwise.
}
\description{
For a given distribution name, such as "norm", checks to see
whether or not p/d/q functions all exist for that distribution --
such as \code{pnorm()}, \code{dnorm()}, and \code{qnorm()}.
Methods for \code{eval_cdf()}, \code{eval_quantile()}, and \code{eval_density/mass()}
suffice as alternatives. Internal function.
}
