% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R
\name{process_ellipsis_for_enframe}
\alias{process_ellipsis_for_enframe}
\title{Process ellipsis for enframe_ functions}
\usage{
process_ellipsis_for_enframe(...)
}
\arguments{
\item{...}{Arguments to be processed.}
}
\value{
A list of distributions input in \code{...}, with \code{NA_real_} wherever
there is an \code{NA} or \code{NULL} entry.
}
\description{
Takes distributions input in the ellipsis of an enframe_ function,
checks them, and returns a list of those distributions, with NA where
there are empty (NULL) distributions.
}
