% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{pull_parameters}
\alias{pull_parameters}
\title{Parameters of a Distribution}
\usage{
pull_parameters(distribution)
}
\arguments{
\item{distribution}{Probability distribution.}
}
\description{
Parameters of a Distribution
}
