% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{natural0}
\alias{natural1}
\alias{integers}
\alias{arithmetic}
\alias{as_discretes}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{discretes}{\code{\link[discretes]{arithmetic}}, \code{\link[discretes]{as_discretes}}, \code{\link[discretes]{integers}}, \code{\link[discretes:integers]{natural0}}, \code{\link[discretes:integers]{natural1}}}
}}

