% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representation_as_function.R
\name{representation_as_function}
\alias{representation_as_function}
\title{Get a distribution's representation as a function}
\usage{
representation_as_function(distribution, representation)
}
\arguments{
\item{distribution}{Distribution to extract a representation from.}

\item{representation}{Character, such as \code{"cdf"}. In general, a suffix
to an \code{eval_} function.}
}
\value{
A function of the representation.
}
\description{
Get a distribution's representation as a function
}
