% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_fmls_to_enframe.R
\name{send_fmls_to_enframe}
\alias{send_fmls_to_enframe}
\title{Send arguments to an \code{enframe_} function}
\usage{
send_fmls_to_enframe(variate, eval_fn)
}
\arguments{
\item{variate}{One of \code{"uni"}, \code{"bi"}, or \code{"multi"}.}

\item{eval_fn}{String; suffix of an \code{eval_} function, specifying the
distributional representation to be evaluated.}
}
\value{
Either the \code{enframe_uni()}, \code{enframe_bi()}, or \code{enframe_multi()}
function (depending on the \code{variate} input), evaluated at the
encapsulating function's arguments, with the \code{eval_fn} argument
also passed.
}
\description{
When used within a (encapsulating) function, \code{send_fmls_to_enframe()}
sends the inputs of the encapsulating function to either a univariate,
bivariate, or multivariate enframer.
}
