% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{set_density}
\alias{set_density}
\title{Specify a distributional representation}
\usage{
set_density(distribution, fun)
}
\arguments{
\item{distribution}{A distribution object.}

\item{fun}{A function specifying the distributional representation,
with arguments only for the distribution variables (not, for example,
parameters).}

\item{env}{Environment; if upon executing the function \code{fun} an object
has not been defined, this environment will be searched,
followed by the global one specified in
the \code{distribution()} function.}
}
\description{
Add a representation to a distribution, such as density, CDF,
hazard function, PMF, etc. If a representation is already specified,
it will be replaced with a warning message.
}
