% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distribution.R
\name{validate_distribution}
\alias{validate_distribution}
\title{Validate a distribution object}
\usage{
validate_distribution(object)
}
\arguments{
\item{object}{Object to be tested}
}
\value{
The input object, invisibly.
}
\description{
A low-level checker that all the fields of a distribution object are in
place. Does not check whether the distributional representations and
properties are valid.
}
