% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Check whether parameter values are valid}
\usage{
validate_parameters(distribution, ..., .stop = FALSE)
}
\arguments{
\item{distribution}{distribution}

\item{...}{Named values to check. Evaluated in caller environment.}

\item{.stop}{Stop if parameters are not valid?}
}
\value{
Single logical indicating whether the parameter values in \code{...}
fall within the parameter space, given the values of parameters that are
already resolved. May return \code{NA} if validity cannot be determined;
for example, if \code{sigma > mu} is a restriction, testing \code{sigma = 4}
will return \code{NA} unless \code{mu} is known.
}
\description{
Check whether parameter values fall within a distribution family's
parameter space.
}
