% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe.R
\name{vec_recycle_common_and_rename}
\alias{vec_recycle_common_and_rename}
\title{vec_recycle_common and fill in missing names}
\usage{
vec_recycle_common_and_rename(...)
}
\arguments{
\item{...}{Vectors to recycle.}
}
\value{
A named list, not necessarily with unique names.
}
\description{
Recycles input vectors to a common length, as in
\code{vctrs::vec_recycle_common()}, but names that were originally unnamed
are given names: when an input is a symbol, that symbol becomes the
component name; otherwise, the name is \code{...i}, where \code{i} is the entry
number.
}
\examples{
a <- 1:3
distionary:::vec_recycle_common_and_rename(a, b = 2, 32 + 5, f = a, f = 4)
distionary:::vec_recycle_common_and_rename(!!!list(a, b = 2, 32 + 5))
}
