% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{atoms}
\alias{atoms}
\alias{continuous_part}
\title{Atomic and Continuous Parts of a Support}
\usage{
atoms(x)

continuous_part(x)
}
\arguments{
\item{x}{A support object or a distribution.}
}
\value{
For \code{atoms()}, a \code{discretes} object. For \code{continuous_part()}, a
two-column numeric matrix of intervals (\code{lower}, \code{upper}).
}
\description{
Extract the atomic (discrete) part or the continuous part of a support. Each
accepts either a support object or a distribution.
}
\examples{
atoms(mixed(atoms = 0, continuous = c(0, Inf)))
continuous_part(continuous(c(0, 1), c(3, 4)))
}
\seealso{
Other Support: 
\code{\link{is_support}()},
\code{\link{support}()},
\code{\link{support-construction}}
}
\concept{Support}
