% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe_cdf.R, R/eval_bi_cdf.R, R/eval_cdf.R,
%   R/eval_multi_cdf.R
\name{enframe_cdf}
\alias{enframe_cdf}
\alias{enframe_bi_cdf}
\alias{enframe_multi_cdf}
\alias{eval_bi_cdf}
\alias{eval_bi_cdf.default}
\alias{eval_cdf}
\alias{eval_multi_cdf}
\title{Cumulative Distribution Function}
\usage{
enframe_cdf(..., at, arg_name = ".arg", fn_prefix = "cdf", sep = "_")

enframe_bi_cdf(..., x, y, arg_name = NULL, fn_prefix = "cdf", sep = "_")

enframe_multi_cdf(..., .l, arg_name = NULL, fn_prefix = "cdf", sep = "_")

eval_bi_cdf(distribution, x, y)

\method{eval_bi_cdf}{default}(distribution, x, y)

eval_cdf(distribution, at)

eval_multi_cdf(distribution, .l)
}
\arguments{
\item{at}{A vector of values to evaluate the distributional
representation at.}

\item{arg_name}{For \code{enframe_}, name(s) of the column containing
the function arguments.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one distribution, the
character that will be separating the \code{fn_name} and the distribution name.}

\item{x, y}{Vectors of values to evaluate the (bivariate) distributional
representation at.}

\item{.l}{A list of vectors, one for each dimension of the distribution,
to evaluate the distributional representation at. Optionally, a single
vector of the same length, for evaluating the representation at a single
point.}

\item{distribution, ...}{A distribution, or possibly multiple
distributions in the case of \code{...}.}
}
\value{
The evaluated cdf in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's cumulative distribution function (cdf).
}
\examples{
d1 <- dst_unif(0, 4)
d2 <- dst_pois(1.1)
eval_cdf(d1, at = 0:4)
enframe_cdf(d1, at = 0:4)
enframe_cdf(d1, d2, at = 0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_chf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
