% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_chf.R
\name{eval_chf}
\alias{eval_chf}
\alias{enframe_chf}
\title{Cumulative Hazard Function}
\usage{
eval_chf(distribution, at)

enframe_chf(..., at, arg_name = ".arg", fn_prefix = "chf", sep = "_")
}
\arguments{
\item{distribution}{A distribution, inheriting the class \code{"dst"}.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}
}
\value{
The evaluated cumulative hazard
in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's cumulative hazard function (chf).
}
\examples{
d <- dst_unif(0, 4)
eval_chf(d, at = 0:4)
enframe_chf(d, at = 0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
