% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_density.R
\name{eval_density}
\alias{eval_density}
\alias{enframe_density}
\alias{enframe_bi_density}
\alias{enframe_density.multi_dst}
\title{Probability Density Function}
\usage{
eval_density(distribution, at)

enframe_density(
  ...,
  at,
  arg_name = ".arg",
  fn_prefix = "density",
  sep = "_",
  strict = TRUE
)

enframe_bi_density(
  ...,
  x,
  y,
  arg_name = NULL,
  fn_prefix = "density",
  sep = "_",
  strict = TRUE
)

enframe_density.multi_dst(
  ...,
  .l,
  arg_name = NULL,
  fn_prefix = "density",
  sep = "_",
  strict = TRUE
)
}
\arguments{
\item{distribution, ...}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{at}{A vector of values to evaluate the distributional
representation at.}

\item{arg_name}{For \code{enframe_}, name(s) of the column containing
the function arguments.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one distribution, the
character that will be separating the \code{fn_name} and the distribution name.}

\item{x, y}{Vectors of values to evaluate the (bivariate) distributional
representation at.}

\item{.l}{A list of vectors, one for each dimension of the distribution,
to evaluate the distributional representation at. Optionally, a single
vector of the same length, for evaluating the representation at a single
point.}
}
\value{
The evaluated density
in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's probability density function (pdf).
}
\examples{
d <- dst_unif(0, 4)
eval_density(d, at = 0:4)
enframe_density(d, at = 0:4)
eval_density(dst_pois(1), at = 0:4, strict = FALSE)
}
\seealso{
Other distributional representations: 
\code{\link{enframe_cdf}()},
\code{\link{eval_chf}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
