% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension.R
\name{dimension}
\alias{dimension}
\title{Dimension of a Distribution}
\usage{
dimension(distribution)
}
\arguments{
\item{distribution}{A distribution, inheriting the class
\code{"dst"}, \code{"bidst"}, or \code{"multidst"}.}
}
\value{
Integer.
}
\description{
Retrieves the number of variables associated with a probability distribution.
}
\examples{
dimension(dst_norm(0, 1))
}
