% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile_from_cdf.R
\name{directional_inverse}
\alias{directional_inverse}
\title{Algorithm to Compute a Directional Inverse}
\usage{
directional_inverse(distribution, p, low, high, tol, maxiter, direction)
}
\arguments{
\item{distribution}{A distribution with a specified cdf.}

\item{p}{Single value for which to calculate the left inverse.}

\item{low, high}{Single numeric values forming a range
within which to search for the solution.}

\item{tol}{Maximum tolerance for the solution.}

\item{maxiter}{Number of iterations to attempt before
quitting, if the tolerance has not been reached. by then.}

\item{direction}{One of \code{"left"} for calculating left-inverse, or
\code{"right"} for calculating right-inverse.}
}
\description{
Calculates the smallest value for which a function
\code{f} evaluates to be greater than or equal to
\code{y} -- that is, the left inverse of \code{f}
at \code{y}. Intended for internal use only.
}
\details{
This algorithm works by progressively
cutting the specified range in half, so that the width
of the range after k iterations is 1/2^k times the
original width.
}
