% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distionary-package.R
\docType{package}
\name{distionary-package}
\alias{distionary}
\alias{distionary-package}
\title{distionary: Create and Evaluate Probability Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create probability distribution objects from a list of common distribution families, or make your own distribution. Evaluate distribution properties, even if it is not specified in the distribution's definition.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://distionary.probaverse.com/}
  \item \url{https://github.com/probaverse/distionary}
  \item Report bugs at \url{https://github.com/probaverse/distionary/issues}
}

}
\author{
\strong{Maintainer}: Vincenzo Coia \email{vincenzo.coia@gmail.com}

Other contributors:
\itemize{
  \item Amogh Joshi [contributor]
  \item Shuyi Tan [contributor]
  \item Zhipeng Zhu [contributor]
}

}
\keyword{internal}
