% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{is_distribution}
\alias{is_distribution}
\alias{is.distribution}
\title{#' Constructor Function for "dst" Objects
#'
#' @param l List containing the components of a distribution object.
#' @param variable Type of random variable: "continuous", "discrete",
#'   or "mixed".
#' @param ... Attributes to add to the list.
#' @param class If making a subclass, specify its name here.
#' @export
new_distribution <- function(l, variable, ...,
class = character()) {
structure(
l,
variable = variable,
class    = c(class, "dst")
)
}}
\usage{
is_distribution(object)

is.distribution(object)
}
\arguments{
\item{object}{Object to be tested}
}
\description{
Distribution Objects
}
