% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R, R/new_distribution.R
\name{distribution}
\alias{distribution}
\alias{is_distribution}
\alias{is.distribution}
\title{Make a distribution from scratch}
\usage{
distribution(
  ...,
  .name = NULL,
  .vtype = NULL,
  .params = params(),
  .parenv = rlang::caller_env()
)

is_distribution(object)

is.distribution(object)
}
\arguments{
\item{...}{Properties of the distribution as name-value pairs. See details
for special property names.}

\item{.vtype}{Type of random variable. Common options
include "continuous", "discrete", or "mixed", but remains open for
extensibility.}

\item{.params}{Vector of unquoted variable names acting as the
distribution's parameters, if relevant.}

\item{.parenv}{Parent environment of objects specified in \code{...}. For
advanced use, as the default option (the caller environment) suffices in
most situations. An example is building a subclass of a distribution, and
using the larger class's representation environment as the parent.}

\item{object}{Object to be tested}
}
\value{
A distribution object with nothing in it.
}
\description{
Initiate a distribution object by specifying the representation space.
}
