% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R, R/new_distribution.R
\name{distribution}
\alias{distribution}
\alias{is_distribution}
\alias{is.distribution}
\title{Make a distribution}
\usage{
distribution(..., .vtype = NULL, .name = NULL)

is_distribution(object)

is.distribution(object)
}
\arguments{
\item{...}{Name-value pairs to add to the distribution list.}

\item{.vtype}{The variable type, indicating the values for which
the distribuion provides probabilities for. For example,.
\code{"continuous"} implies that the distribution acts on real numbers;
\code{"discrete"} implies discrete values.}

\item{.name}{If this distribution has a name (e.g., "Normal"),
optionally specify it here.}

\item{object}{Object to be tested}
}
\value{
A distribution object.
}
\description{
Make a distribution object by specifying its representations
(e.g., cdf, density, mean, etc.).
Some representations, if not included, will be calculated based
on other representations that are included (e.g., quantile from cdf).
A list of these representations can be found in the details.
Note that any object can be added to the distribution.
}
\details{
Currently, at least two representations are required to be specified if
you want to be able to retrieve other properties without specifying them:
\itemize{
\item \code{cdf}: the cumulative distribution function.
\item \code{density}, the probability density function for continuous variables,
or \code{pmf}, the probability mass function for discrete variables.
}

Other representations that will be retrieved if missing include:
\itemize{
\item \code{survival}: the survival function, or one minus the cdf.
\item \code{hazard}: the hazard function, for continuous variables only.
\item \code{chf}: the cumulative hazard function, for continuous variables only.
\item \code{quantile}: the quantile function, or left-inverse of the cdf.
\item \code{realise} or \code{realize}: a function that takes an integer and generates
a vector of that many random draws from the distribution.
\item \code{odds}: for discrete variables, the probability odds function
(pmf / (1 - pmf))
\item \code{return}: the quantiles associated with the provided return periods,
where events are exceedances.
}

All functions should be vectorized.

Properties that can be retrieved if not specified include the following:
\itemize{
\item \code{mean}, \code{stdev}, \code{variance}, \code{skewness}, \code{median} are self-explanatory.
\item \code{kurtosis_exc} and \code{kurtosis} are the distribution's excess
kurtosis and regular kurtosis.
\item \code{range}: A vector of the minimum and maximum value of a distribution's
support.
}

A later version of distionary will allow for the specification of
custom representations / properties.
}
