% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distribution.R
\name{is_distribution}
\alias{is_distribution}
\alias{is.distribution}
\title{Distribution Objects}
\usage{
is_distribution(object)

is.distribution(object)
}
\arguments{
\item{object}{Object to be tested}
}
\description{
Distribution Objects
}
