% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_bern.R
\name{dst_bern}
\alias{dst_bern}
\title{Bernoulli Distribution}
\usage{
dst_bern(p)
}
\arguments{
\item{p}{}
}
\description{
Makes a distribution belonging to the family of
bernoulli distributions.
}
\examples{
dst_bern(0.3)

}
