% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_binom.R
\name{dst_binom}
\alias{dst_binom}
\title{Binomial Distribution}
\usage{
dst_binom(size, p)
}
\arguments{
\item{size}{number of trials}

\item{p}{success probability for each trial}
}
\description{
Makes a distribution belonging to the family of
binomial distributions.
}
\examples{
dst_binom(10, 0.6)

}
