% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_cauchy.R
\name{dst_cauchy}
\alias{dst_cauchy}
\title{Cauchy Distribution}
\usage{
dst_cauchy(location, scale)
}
\arguments{
\item{location}{Location parameter}

\item{scale}{0}
}
\description{
Makes a distribution belonging to the family of
cauchy distributions.
}
\examples{
dst_cauchy(0, 1)

}
