% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_chisq.R
\name{dst_chisq}
\alias{dst_chisq}
\title{Chi-squared Distribution}
\usage{
dst_chisq(df)
}
\arguments{
\item{df}{degrees of freedom}
}
\description{
Makes a distribution belonging to the family of
Chi-squared distributions.
}
\examples{
dst_chisq(3)

}
