% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_exp.R
\name{dst_exp}
\alias{dst_exp}
\title{Exponential Distribution}
\usage{
dst_exp(rate)
}
\arguments{
\item{rate}{Rate.}
}
\description{
Makes a distribution belonging to the family of
exponential distributions.
}
\examples{
dst_exp(1)

}
