% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_f.R
\name{dst_f}
\alias{dst_f}
\title{F Distribution}
\usage{
dst_f(df1, df2)
}
\arguments{
\item{df1}{degree of freedom, positive.}

\item{df2}{degree of freedom, positive.}
}
\description{
Makes a distribution belonging to the family of
F distributions.
}
\examples{
dst_f(2, 3)
}
