% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_gamma.R
\name{dst_gamma}
\alias{dst_gamma}
\title{Gamma Distribution}
\usage{
dst_gamma(alpha, beta)
}
\arguments{
\item{alpha}{shape > 0}

\item{beta}{rate > 0}
}
\description{
Makes a distribution belonging to the family of
Gamma distributions.
}
\examples{
dst_gamma(2, 1)

}
