% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_geom.R
\name{dst_geom}
\alias{dst_geom}
\title{Geometric Distribution}
\usage{
dst_geom(p)
}
\arguments{
\item{p}{probability of success in each trial. 0 < p <= 1}
}
\description{
Makes a distribution belonging to the family of
geometric distributions.
}
