% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_geom.R
\name{dst_geom}
\alias{dst_geom}
\title{Geometric Distribution}
\usage{
dst_geom(prob)
}
\arguments{
\item{prob}{Probability of success in each trial; 0 < \code{p} <= 1.}
}
\description{
Makes a distribution belonging to the family of
geometric distributions.
}
\examples{
dst_geom(0.4)
}
