% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_gev.R
\name{dst_gev}
\alias{dst_gev}
\title{Generalized Extreme Values Distribution}
\usage{
dst_gev(location, scale, shape)
}
\arguments{
\item{location}{real number}

\item{scale}{positive}

\item{shape}{real number}
}
\description{
Makes a distribution belonging to the family of
gev distributions.
}
\examples{
dst_gev(0, 1, 1)


}
