% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_gev.R
\name{dst_gev}
\alias{dst_gev}
\title{Generalized Extreme Value Distribution}
\usage{
dst_gev(location, scale, shape)
}
\arguments{
\item{location}{Location parameter; numeric.}

\item{scale}{Scale parameter; positive numeric.}

\item{shape}{Shape parameter; numeric.}
}
\description{
Makes a distribution belonging to the family of
Generalized Extreme Value (GEV) distributions.
}
\examples{
dst_gev(0, 1, 1)
}
