% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_hyper.R
\name{dst_hyper}
\alias{dst_hyper}
\title{Hypergeometric Distribution}
\usage{
dst_hyper(m, n, k)
}
\arguments{
\item{m}{The number of red balls in the urn.}

\item{n}{The number of non-red balls in the urn.}

\item{k}{the number of balls drawn from the urn (between 0 and \code{m + n}).}
}
\value{
A Hypergeometric distribution.
}
\description{
Makes a distribution belonging to the family of
hypergeometric distributions. This distribution represents the
number of red balls drawn from an urn of balls.
}
\examples{
dst_hyper(1, 5, 10)

}
