% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_hyper.R
\name{dst_hyper}
\alias{dst_hyper}
\title{Hypergeometric Distribution}
\usage{
dst_hyper(K, N, n)
}
\arguments{
\item{K}{the number of success states in the population}

\item{N}{the population size}

\item{n}{the number of draws}
}
\description{
Makes a distribution belonging to the family of
hypergeometric distributions.
}
\examples{
dst_hyper(1, 5, 10)

}
