% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_lnorm.R
\name{dst_lnorm}
\alias{dst_lnorm}
\title{Log Normal Distribution}
\usage{
dst_lnorm(meanlog, sdlog)
}
\arguments{
\item{meanlog, sdlog}{Mean and variance of the distribution
on a log scale.}
}
\value{
A Log Normal distribution.
dst_lnorm(0, 1)
}
\description{
Makes a distribution belonging to the family of
Log Normal distributions.
}
