% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_nbinom.R
\name{dst_nbinom}
\alias{dst_nbinom}
\title{Negative binomial Distribution}
\usage{
dst_nbinom(size, prob)
}
\arguments{
\item{size}{Target for number of successful trials; non-negative.}

\item{prob}{Probability parameter; between 0 and 1.}
}
\description{
Makes a distribution belonging to the family of
negative binomial distributions.
}
\examples{
dst_nbinom(10, 0.5)
}
