% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_norm.R
\name{dst_norm}
\alias{dst_norm}
\title{Normal (Gaussian) Distribution}
\usage{
dst_norm(mean, sd)
}
\arguments{
\item{mean, sd}{Mean and standard deviation of the distribution.}
}
\value{
A Normal distribution.
}
\description{
Makes a Normal (Gaussian) distribution.
}
\examples{
dst_norm(0, 1)
}
