% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_null.R
\name{dst_null}
\alias{dst_null}
\title{Empty / NULL distribution}
\usage{
dst_null()
}
\value{
A null distribution, having no properties defined.
}
\description{
Sometimes it's convenient to work with a distribution object that is empty,
akin to a \code{NULL} value. This is especially true when programmatically
outputting distributions, such as when a distribution fails to fit to
data. This constructor makes such a distribution object.
It always evaluates to an \code{NA} vector.
}
\examples{
x <- dst_null()
mean(x)
eval_pmf(x, at = 1:10)
enframe_hazard(x, at = 1:10)
}
