% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_null.R
\name{dst_null}
\alias{dst_null}
\title{Null Distribution}
\usage{
dst_null()
}
\value{
A Null distribution.
}
\description{
Sometimes it's convenient to work with a distribution object that is
akin to a missing value. This is especially true when programmatically
outputting distributions, such as when a distribution fails to fit to
data. This function makes such a distribution object. It always evaluates
to \code{NA}.
}
\examples{
x <- dst_null()
mean(x)
eval_pmf(x, at = 1:10)
enframe_hazard(x, at = 1:10)
}
