% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_pearson3.R
\name{dst_pearson3}
\alias{dst_pearson3}
\title{Pearson Type III distribution}
\usage{
dst_pearson3(location, scale, shape)
}
\arguments{
\item{location}{Location parameter, specifying how to shift the
Gamma distribution; single numeric.}

\item{scale}{Scale parameter of the Gamma distribution;
single positive numeric.}

\item{shape}{Shape parameter of the Gamma distribution;
single positive numeric.}
}
\value{
A Pearson Type III distribution.
}
\description{
Makes a Pearson Type III distribution, which is a Gamma distribution,
but shifted.
}
\examples{
dst_pearson3(1, 1, 1)
}
