% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_pearson3.R
\name{dst_pearson3}
\alias{dst_pearson3}
\title{Pearson Type III distribution}
\usage{
dst_pearson3(location, scale, shape)
}
\arguments{
\item{location, scale, shape}{Distribution parameters. Need scale > 0
and shape > 0.}
}
\value{
A Log Pearson Type III distribution.
}
\description{
Pearson Type III distribution
}
\details{
This is the equivalent of the Gamma distribution in the
stats package with parameters \code{scale} and \code{shape},
}
