% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_t.R
\name{dst_t}
\alias{dst_t}
\title{Student t Distribution}
\usage{
dst_t(df)
}
\arguments{
\item{df}{Degrees of freedom; single positive numeric.}
}
\value{
A Student t distribution.
}
\description{
Makes a Student t distribution.
}
\examples{
dst_t(3)

}
