% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_t.R
\name{dst_t}
\alias{dst_t}
\title{The Student t Distribution}
\usage{
dst_t(df)
}
\arguments{
\item{df}{degrees of freedom; positive}
}
\description{
Makes a distribution belonging to the family of
t distributions.
}
\examples{
dst_t(3)

}
