% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_weibull.R
\name{dst_weibull}
\alias{dst_weibull}
\title{Weibull Distribution}
\usage{
dst_weibull(shape, scale)
}
\arguments{
\item{shape}{Shape parameter; positive.}

\item{scale}{Scale parameter; positive.}
}
\description{
Makes a distribution belonging to the family of
Weibull distributions.
}
\examples{
dst_weibull(1, 1)

}
