% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dst_weibull.R
\name{dst_weibull}
\alias{dst_weibull}
\title{Weibull Distribution}
\usage{
dst_weibull(lambda, gamma)
}
\arguments{
\item{lambda}{0 scale}

\item{gamma}{0 shape}
}
\description{
Makes a distribution belonging to the family of
Weibull distributions.
}
