% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile_from_cdf.R
\name{encapsulate_p}
\alias{encapsulate_p}
\title{Find a range of possible outcomes}
\usage{
encapsulate_p(distribution, p, direction)
}
\arguments{
\item{distribution}{A distribution having access to a cdf.}

\item{p}{Vector of values between 0 and 1 (inclusive).}

\item{direction}{One of \code{"left"} for calculating left-inverse, or
\code{"right"} for calculating right-inverse.}
}
\description{
Find a range of possible outcomes where the cdf evaluates to values
that encapsulate the range of probabilities contained in the vector \code{p}.
}
\note{
If 0 or 1 are included in the vector \code{p}, one of the endpoints might
be infinite.
}
