% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-quantile.R
\name{encapsulate_p}
\alias{encapsulate_p}
\title{Find a range of possible outcomes}
\usage{
encapsulate_p(distribution, p)
}
\description{
Find a range of possible outcomes where the cdf evaluates to values
that encapsulate the range of probabilities contained in the vector \code{p}.
}
