% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile_from_cdf.R
\name{encapsulate_p}
\alias{encapsulate_p}
\title{Find a range of possible outcomes}
\usage{
encapsulate_p(distribution, p, strict_inequality = c(TRUE, FALSE))
}
\arguments{
\item{distribution}{Distribution.}

\item{p}{Vector of values between 0 and 1 (inclusive).}

\item{strict_inequality}{Logical vector of length two.
Should the lower bound evaluate to be strictly less than \code{p}?
Should the upper bound evaluate to be strictly greater than \code{p}?
\code{c(TRUE, FALSE)} is the default, useful for calculating left-inverses;
\code{c(FALSE, TRUE)} is useful for calculating right-inverses.}
}
\description{
Find a range of possible outcomes where the cdf evaluates to values
that encapsulate the range of probabilities contained in the vector \code{p}.
}
\note{
If 0 or 1 are included in the vector \code{p}, one of the endpoints might
be infinite.
}
