% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe_variate.R
\name{enframe_uni}
\alias{enframe_uni}
\alias{enframe_bi}
\alias{enframe_multi}
\title{Standardize arguments and enframe}
\usage{
enframe_uni(..., at, arg_name, fn_prefix, sep, eval_fn)

enframe_bi(..., x, y, arg_name, fn_prefix, sep, eval_fn)

enframe_multi(..., .l, arg_name, fn_prefix, sep, eval_fn)
}
\arguments{
\item{eval_fn}{Character name of the eval function to enframe;
e.g., \code{"eval_multi_cdf"}.}
}
\value{
Calls the \code{enframe_general()} function with a standardized
\code{.l} input.
}
\description{
Internal. An intermediate step in the \code{enframe_}
family of functions, these functions
standardize the arguments in which to evaluate the distributional
representation at, before calling a general enframer function.
}
\details{
\itemize{
\item The univariate version puts the \code{at} input into a list,
named by \code{arg_name}.
\item The bivariate version recycles \code{x} and \code{y} into a common length,
puts them in a list, and names them by \code{arg_name} if present.
\item The multivariate version recycles the \code{.l} input, after making it
a list in the case of a vector input (representing a tuple).
}
}
