% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enframe_general.R
\name{name_distributions}
\alias{name_distributions}
\title{Column names for enframe functions}
\usage{
name_distributions(..., fn_prefix)
}
\arguments{
\item{...}{f}

\item{fn_prefix}{Name of the prefix to appear in the function
evaluation column(s).}
}
\description{
Derive names for the columns in the enframe_ family of functions;
"input" being the function argument being evaluated, and
"output" being the evaluated function.
}
