% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_bind_within.R
\name{env_bind_representations}
\alias{env_bind_representations}
\title{First binds representations as expressions with \code{env_bind_exprs()},
then evaluates the expression within the representation environment
if the expression evaluates to a function.}
\usage{
env_bind_representations(.env, ...)
}
\description{
First binds representations as expressions with \code{env_bind_exprs()},
then evaluates the expression within the representation environment
if the expression evaluates to a function.
}
