% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile_from_cdf.R
\name{eval_quantile_from_cdf}
\alias{eval_quantile_from_cdf}
\title{Evaluate Quantiles from a CDF}
\usage{
eval_quantile_from_cdf(distribution, at, tol, maxiter)
}
\arguments{
\item{distribution}{A distribution having access to a cdf.}

\item{at}{A vector of values for which to evaluate the quantile function.}

\item{tol, maxiter}{Tolerance (a small positive number) and maximum number
of iterations}
}
\description{
Intended for internal use only. Calls the \code{directional_inverse()}
algorithm for each new entry of \code{at}.
}
