% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile_from_network.R
\name{eval_quantile_from_network}
\alias{eval_quantile_from_network}
\title{Evaluate Quantiles from a CDF}
\usage{
eval_quantile_from_network(distribution, at, tol = 1e-09, maxiter = 200)
}
\arguments{
\item{distribution}{A distribution having access to a cdf.}

\item{at}{A vector of values for which to evaluate the quantile function.}

\item{tol, maxiter}{Tolerance (a small positive number) and maximum number
of iterations}
}
\value{
The \code{at}-quantiles of the distribution.
}
\description{
Pulls together all the pieces needed to calculate the left inverse
of the CDF. Intended for internal use only. Calls the
\code{encapsulate_p()} function to figure out where to start looking for
solutions, then \code{directional_inverse()} to run the algorithm.
}
