% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdqr_gev.R
\name{gev_range}
\alias{gev_range}
\title{Range of a GEV distribution}
\usage{
gev_range(location, scale, shape)
}
\arguments{
\item{location}{Location parameter; numeric.}

\item{scale}{Scale parameter; positive numeric.}

\item{shape}{Shape parameter; numeric.}
}
\value{
A vector of length 2 indicating the support of the distribution.
}
\description{
Range of a GEV distribution
}
