% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations-dst_gev.R
\name{pgev}
\alias{pgev}
\alias{qgev}
\alias{dgev}
\title{#' @export
eval_cdf.gev <- function(distribution, at) {
r <- range(distribution)
with(parameters(distribution), {
t <- gev_t_function(at, location = location, scale = scale, shape = shape)
res <- exp(-t)
res[at < r\link{1L}] <- 0
res[at > r\link{2L}] <- 1
res
})
}
Representations of the Generalized Extreme Value Distribution}
\usage{
pgev(q, location, scale, shape)

qgev(p, location, scale, shape)

dgev(x, location, scale, shape)
}
\description{
#' @export
eval_cdf.gev <- function(distribution, at) {
r <- range(distribution)
with(parameters(distribution), {
t <- gev_t_function(at, location = location, scale = scale, shape = shape)
res <- exp(-t)
res[at < r\link{1L}] <- 0
res[at > r\link{2L}] <- 1
res
})
}
Representations of the Generalized Extreme Value Distribution

#' @export
eval_quantile.gev <- function(distribution, at) {
with(parameters(distribution), {
invalid_at <- at < 0 | at > 1
if (shape == 0) {
res <- location - scale * log(-log(at))
} else {
res <- location + scale * ((-log(at))^(-shape) - 1) / shape
}
res\link{invalid_at} <- NaN
res
})
}

#' @export
eval_density.gev <- function(distribution, at, strict = TRUE) {
r <- range(distribution)
with(parameters(distribution), {
t <- gev_t_function(at, location = location, scale = scale, shape = shape)
res <- t^(shape + 1) / scale * exp(-t)
res[at < r\link{1L}] <- 0
res[at > r\link{2L}] <- 0
res
})
}
}
