% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representations-dst_gev.R
\name{gev_t_function}
\alias{gev_t_function}
\title{'t()' function for calculating GEV quantities}
\usage{
gev_t_function(x, location, scale, shape)
}
\arguments{
\item{x}{Argument of the function; vectorized.}

\item{location}{Location parameter; numeric.}

\item{scale}{Scale parameter; positive numeric.}

\item{shape}{Shape parameter; numeric.}
}
\description{
't()' function for calculating GEV quantities
}
\note{
The shape parameter is not vectorized. This function is only
intended to be used when location, scale, and shape are scalars.
}
\seealso{
See the Wikipedia entry for the GEV distribution,
https://en.wikipedia.org/wiki/Generalized_extreme_value_distribution
}
