% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hazard.R
\name{eval_hazard}
\alias{eval_hazard}
\alias{enframe_hazard}
\title{Hazard Function}
\usage{
eval_hazard(distribution, at)

enframe_hazard(..., at, arg_name = ".arg", fn_prefix = "hazard", sep = "_")
}
\arguments{
\item{distribution}{A distribution, inheriting the class \code{"dst"}.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}
}
\value{
The evaluated hazard in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's hazard function.
}
\examples{
d <- dst_unif(0, 4)
eval_hazard(d, at = 0:4)
enframe_hazard(d, at = 0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_chf}()},
\code{\link{eval_odds}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
