% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{is_support}
\alias{is_support}
\title{Test for a Support Object}
\usage{
is_support(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} is a support object, otherwise \code{FALSE}.
}
\description{
Test for a Support Object
}
\seealso{
Other Support: 
\code{\link{atoms}()},
\code{\link{support}()},
\code{\link{support-construction}}
}
\concept{Support}
