% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdq_lp3.R
\name{plp3}
\alias{plp3}
\alias{dlp3}
\alias{qlp3}
\alias{rlp3}
\title{Representations of the Log Pearson Type III Distribution}
\usage{
plp3(x, meanlog, sdlog, skew, lower.tail = TRUE)

dlp3(x, meanlog, sdlog, skew)

qlp3(p, meanlog, sdlog, skew)

rlp3(n, meanlog, sdlog, skew)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{meanlog}{Parameter representing the mean of the random variable
in log (base e) space; numeric. Vectors are allowed except for \code{rlp3()}.}

\item{sdlog}{Parameter representing the standard deviation of the
random variable in log (base e) space; positive numeric.
Vectors are allowed except for \code{rlp3()}.}

\item{skew}{Parameter representing the skewness of the random variable
in log (base e) space; positive numeric.
Vectors are allowed except for \code{rlp3()}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are
\code{P(X <= x)}, otherwise, \code{P(X > x)}.}

\item{p}{Vector of probabilities.}

\item{n}{Single positive whole number; number of observations to draw
from the distribution.}
}
\value{
Vector of evaluated LP3 distribution, with length
equal to the recycled lengths of \code{x}/\code{p}, \code{meanlog}, \code{sdlog}, and
\code{skew}. For \code{rlp3()}, a vector of length \code{n}.
}
\description{
Representations of the Log Pearson Type III (LP3) Distribution
}
\examples{
plp3(1:10, meanlog = 0, sdlog = 1, skew = 1)
dlp3(1:10, meanlog = 1:10, sdlog = 2, skew = 0)
qlp3(1:9 / 10, meanlog = 2, sdlog = 10, skew = 2)
set.seed(1)
rlp3(10, meanlog = 2, sdlog = 10, skew = 2)
}
