% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R, R/mean.R, R/median.R, R/skewness.R,
%   R/stdev.R, R/variance.R
\name{kurtosis_raw}
\alias{kurtosis_raw}
\alias{kurtosis_exc}
\alias{mean.dst}
\alias{median.dst}
\alias{skewness}
\alias{stdev}
\alias{variance}
\title{Moments of a Distribution}
\usage{
kurtosis_raw(distribution, ...)

kurtosis_exc(distribution, ...)

\method{mean}{dst}(x, ...)

\method{median}{dst}(x, ...)

skewness(distribution, ...)

stdev(distribution, ...)

variance(distribution, ...)
}
\arguments{
\item{...}{Other arguments to pass to \code{integrate()} when
deploying numerical algorithm.}

\item{x, distribution}{Distribution to compute moment of.}
}
\value{
A single numeric.
}
\description{
Get common moment-related quantities of a
distribution.
}
\details{
If there is no moment specified in the distribution
definition, then moments are calculated numerically using
\code{stats::integrate()}.

Median is calculated as the 0.5-quantile. So, when the median is
non-unique, we take the smallest of the possibilities.
}
\examples{
a <- dst_gpd(0, 1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
mean(a)
variance(b)
kurtosis_raw(c)
kurtosis_exc(c)
}
