% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R, R/mean.R, R/median.R, R/skewness.R,
%   R/stdev.R, R/variance.R
\name{kurtosis}
\alias{kurtosis}
\alias{kurtosis_exc}
\alias{mean.dst}
\alias{median.dst}
\alias{skewness}
\alias{stdev}
\alias{variance}
\title{Moments of a Distribution}
\usage{
kurtosis(distribution)

kurtosis_exc(distribution)

\method{mean}{dst}(x, ...)

\method{median}{dst}(x, ...)

skewness(distribution)

stdev(distribution)

variance(distribution)
}
\arguments{
\item{distribution}{Distribution to compute skewness from.}

\item{x}{Distribution to calculate median from.}

\item{...}{Not used.}
}
\value{
A single numeric.
}
\description{
Get common moment-related quantities of a
distribution: mean, variance, standard deviation (stdev),
skewness, and kurtosis. If these quantities are not supplied in the
distribution's definition, they will be calculated numerically.
}
\details{
If there is no method associated with a subclass of
\code{x}, then moments are calculated using
\code{stats::integrate()} from the quantile function.

Median is calculated as the 0.5-quantile when not found in the
distribution. So, when the median is
non-unique, the smallest of the possibilities is taken.
}
\note{
Beware that if a quantity is being calculated numerically
for a non-continuous (e.g., discrete) distribution, the calculation
could be highly approximate. An upcoming version of distionary will
resolve this issue.
}
\examples{
a <- dst_gpd(1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
mean(a)
variance(b)
kurtosis(c)
kurtosis_exc(c)
}
