% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.R, R/mean.R, R/median.R, R/skewness.R,
%   R/stdev.R, R/variance.R
\name{kurtosis_raw}
\alias{kurtosis_raw}
\alias{kurtosis_exc}
\alias{mean.dst}
\alias{median.dst}
\alias{skewness}
\alias{stdev}
\alias{stdev.dst}
\alias{variance}
\title{Moments of a Distribution}
\usage{
kurtosis_raw(distribution)

kurtosis_exc(distribution)

\method{mean}{dst}(x, ...)

\method{median}{dst}(x, ...)

skewness(distribution)

stdev(distribution)

\method{stdev}{dst}(distribution)

variance(distribution)
}
\arguments{
\item{distribution}{Distribution to compute skewness from.}

\item{x}{Distribution to calculate median from.}

\item{...}{When calculating the mean via integration of the quantile
function, arguments passed to \code{stats::integrate()}.}
}
\value{
A single numeric.
}
\description{
Get common moment-related quantities of a
distribution: mean, variance, standard deviation (sd),
skewness, and kurtosis.
}
\details{
If there is no method associated with a subclass of
\code{x}, then moments are calculated using
\code{stats::integrate()} from the quantile function.

Median is calculated as the 0.5-quantile. So, when the median is
non-unique, we take the smallest of the possibilities.
}
\examples{
a <- dst_gpd(0, 1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
mean(a)
variance(b)
kurtosis_raw(c)
kurtosis_exc(c)
}
