% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile_from_cdf.R
\name{narrow_by_discretes}
\alias{narrow_by_discretes}
\title{Narrow a range of possible values by a discrete}
\usage{
narrow_by_discretes(distribution, p, low, high, discrete, direction)
}
\arguments{
\item{distribution}{A distribution having access to a cdf.}

\item{p}{Value of the cdf to calculate inverse at.}

\item{low, high}{Single numerics specifying the lower and upper bound
to look between.}

\item{discrete}{Numeric value indicating a discrete point to possibly
narrow the range by. Could be \code{NA}; could have length 0; could be
outside of the range \code{c(low, high)}.}

\item{direction}{One of \code{"left"} for calculating left-inverse, or
\code{"right"} for calculating right-inverse.}
}
\description{
When finding the inverse of the cdf at \code{p}, and with the solution
between \code{low} and \code{high}, this function narrows this range based
on an intermediate discrete value, and zeroes-in on that
discrete value if it's the solution.
}
\note{
The benefit of this step when inverting a cdf is to return
the exact discrete value if \code{p} falls within its jump discontinuity.
}
