% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distribution.R
\name{new_distribution}
\alias{new_distribution}
\title{Constructor Function for "dst" Objects}
\usage{
new_distribution(l, ..., class = character())
}
\arguments{
\item{l}{Object containing the components of a distribution object.}

\item{...}{Attributes to add to the object.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor Function for "dst" Objects
}
