% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distribution.R
\name{new_dst}
\alias{new_dst}
\alias{new_bidst}
\alias{new_multidst}
\alias{new_distribution}
\title{Constructor functions for distribution objects}
\usage{
new_dst(l, variable, ..., class = character())

new_bidst(l, variable, ..., class = character())

new_multidst(l, variable, dimension, ..., class = character())

new_distribution(l, variable, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a distribution object.}

\item{variable}{Vector of random variable types, one for each dimension.
Possibilities are "continuous", "discrete", or "mixed".}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}

\item{dimension}{Integer; number of variables associated with the
distribution.}
}
\description{
Create an object of class \code{"dst"} (univariate), \code{"bidst"} (bivariate),
or \code{"multidst"} (multivariate).
}
\note{
\code{new_distribution()} is included for legacy purposes, and is
identical to \code{new_dst()}.
}
