% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_distribution.R
\name{new_distribution}
\alias{new_distribution}
\title{Constructor Function for Distribution Objects}
\usage{
new_distribution(l, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a distribution object.}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\value{
The original list \code{l} with the specified attributes.
}
\description{
Low level constructor function for creating distribution objects.
Consider the user-facing constructor function, \code{distribution()}.
}
\examples{
new_distribution(list(mean = 3))
}
