% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_enframe.R
\name{new_enframe_uni}
\alias{new_enframe_uni}
\alias{new_enframe_bi}
\alias{new_enframe_multi}
\title{Make a new \code{enframe_} function}
\usage{
new_enframe_uni(repres)

new_enframe_bi(repres)

new_enframe_multi(repres)
}
\arguments{
\item{repres}{String; name of the representation, such as \code{"density"}
or \code{"cdf"}.}
}
\value{
A function that evaluates the specified representation of inputted
distributions, and places the results in a data frame.
\itemize{
\item The \verb{_uni} flavour provides the \code{at} argument to place a single
vector to evaluate the representation, as in \code{enframe_cdf()}.
\item The \verb{_bi} flavour provides the \code{x} and \code{y} arguments to hold
vectors to evaluate the representation, as in \code{enframe_bi_cdf()}.
\item The \verb{_multi} flavour provides the \code{.l} argument to hold a list of
vectors to evaluate the representation, as in \code{enframe_multi_cdf()}.
}
}
\description{
For a given distributional representation, these functions create
an \code{enframe_} function for that representation.
}
