% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_odds.R
\name{eval_odds}
\alias{eval_odds}
\alias{enframe_odds}
\title{Odds Function}
\usage{
eval_odds(distribution, ...)

enframe_odds(...)
}
\arguments{
\item{distribution, ...}{A distribution, or possibly multiple
distributions in the case of \code{...}.}
}
\value{
The evaluated odds in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's odds function. The odds of an event having
probability \code{p} is \code{p / (1 - p)}.
}
\examples{
d <- dst_pois(1)
eval_odds(d, at = c(1, 2, 2.5))
enframe_odds(d, at = 0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_bicdf}()},
\code{\link{eval_cdf}()},
\code{\link{eval_chf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
