% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdq_pearson3.R
\name{ppearson3}
\alias{ppearson3}
\alias{dpearson3}
\alias{qpearson3}
\alias{rpearson3}
\title{Representations of the Pearson Type III Distribution}
\usage{
ppearson3(x, location, scale, shape, lower.tail = TRUE)

dpearson3(x, location, scale, shape, log = FALSE)

qpearson3(p, location, scale, shape)

rpearson3(n, location, scale, shape)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{location}{Parameter representing the left-endpoint of the distribution;
numeric. Vectors are allowed except for \code{rpearson3()}.}

\item{scale}{Scale parameter; positive numeric.
Vectors are allowed except for \code{rpearson3()}.}

\item{shape}{Shape parameter; positive numeric.
Vectors are allowed except for \code{rpearson3()}.}

\item{lower.tail}{Logical; if \code{TRUE} (default), probabilities are
\code{P(X <= x)}, otherwise, \code{P(X > x)}.}

\item{log}{Logical; if \code{TRUE}, probabilities are given as log-probabilities.}

\item{p}{Vector of probabilities.}

\item{n}{Single positive whole number; number of observations to draw
from the distribution.}
}
\value{
Vector of evaluated Pearson Type III distribution, with length
equal to the recycled lengths of \code{x}/\code{p}, \code{location}, \code{scale}, and
\code{shape}. For \code{rpearson3()}, a vector of length \code{n}.
}
\description{
Representations of the Pearson Type III Distribution
}
\examples{
ppearson3(1:10, location = 0, scale = 1, shape = 1)
dpearson3(1:10, location = 1:10, scale = 2, shape = 0)
qpearson3(1:9 / 10, location = -2, scale = 10, shape = 2)
set.seed(1)
rpearson3(10, location = 2, scale = 10, shape = 2)
}
