% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_name.R
\name{pretty_name}
\alias{pretty_name}
\title{Distribution name}
\usage{
pretty_name(distribution, param_digits = 0)
}
\arguments{
\item{distribution}{Distribution object.}

\item{param_digits}{How many significant digits to include when displaying
the parameters? \code{0} if you don't want to display parameters. Length 1
vector.}
}
\value{
A character containing the distribution's name, possibly
followed by parameters in brackets.
}
\description{
Print the name of a distribution, possibly with parameters.
}
\examples{
d <- dst_norm(0.3552, 1.1453)
pretty_name(d)
pretty_name(d, 2)

}
