% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_quantile.R
\name{eval_quantile}
\alias{eval_quantile}
\alias{enframe_quantile}
\title{Distribution Quantiles}
\usage{
eval_quantile(distribution, at)

enframe_quantile(..., at, arg_name = ".arg", fn_prefix = "quantile", sep = "_")
}
\arguments{
\item{distribution}{A distribution, inheriting the class \code{"dst"}.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}
}
\value{
The evaluated quantiles in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's quantiles.
}
\examples{
d <- dst_unif(0, 4)
eval_quantile(d, at = 1:9 / 10)
enframe_quantile(d, at = 1:9 / 10)
}
\seealso{
Other distributional representations: 
\code{\link{eval_chf}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_return}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
