% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.r
\name{range.dst}
\alias{range.dst}
\alias{range.support}
\title{Range of Distribution}
\usage{
\method{range}{dst}(distribution, ...)

\method{range}{support}(support, ...)
}
\arguments{
\item{distribution}{Distribution to compute range from.}

\item{...}{Not used; vestige of the \code{base::range()} S3 generic.}

\item{support}{A support object.}
}
\value{
Vector of length two, containing the minimum and maximum
values of a distribution.
}
\description{
Range returns a vector of length two, with the minimum and maximum
values of the (support of the) distribution.

The \code{support} method gives the smallest and largest values the support
reaches --- its two outermost points, taking the atoms and the continuous
intervals together. Gaps in between are not represented. For the empty
support, both are \code{NA}.
}
\details{
If there are no methods for the distribution's class,
the range is calculated
using \code{eval_quantile()} at 0 and at 1.
}
\examples{
a <- dst_gp(1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
range(a)
range(b)
range(c)
range(continuous(c(0, 1), c(3, 4)))
range(mixed(atoms = -1, continuous = c(0, Inf)))
range(empty_support())
}
