% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities-finite.R, R/range.r
\name{range.finite}
\alias{range.finite}
\alias{range.dst}
\title{Range of Distribution}
\usage{
\method{range}{finite}(..., na.rm = FALSE)

\method{range}{dst}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Single distribution to compute range from.}

\item{na.rm}{Not used; vestige of the \code{base::range()} S3 generic.}
}
\description{
Range returns a 2 index vector with the 0th index
containing the minimum value, and the 1st index containing the maximum value
for a given distribution.
}
\details{
If there are no methods for the distribution's class,
the range is calculated
using \code{eval_quantile()} at 0 and at 1.
}
\examples{
a <- dst_gpd(0, 1, 0.5)
b <- dst_unif(0, 1)
c <- dst_norm(3, 4)
range(a)
range(b)
range(c)
}
