% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_return.R
\name{eval_return}
\alias{eval_return}
\alias{enframe_return}
\title{Return Level Function}
\usage{
eval_return(distribution, at)

enframe_return(..., at, arg_name = ".arg", fn_prefix = "return", sep = "_")
}
\arguments{
\item{distribution}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{at}{Vector of return periods >=1.}

\item{...}{A distribution, or possibly multiple
distributions in the case of \code{...}.}

\item{arg_name}{For \code{enframe_}, name of the column containing
the function arguments.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one distribution, the
character that will be separating the \code{fn_name} and the distribution name.}
}
\value{
The return levels associated with the specified return periods
in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}). This function is simply the quantile
function evaluated at \code{1 - 1 / at}.
}
\description{
Compute return levels (quantiles) from a distribution by inputting
return periods.
}
\examples{
d <- dst_unif(0, 4)
eval_return(d, at = c(2, 25, 100, 200))
enframe_chf(d, at = c(2, 25, 100, 200))
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_chf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
