% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_survival.R
\name{eval_survival}
\alias{eval_survival}
\alias{enframe_survival}
\title{Survival Function}
\usage{
eval_survival(distribution, ...)

enframe_survival(..., at, arg_name = ".arg", fn_prefix = "survival", sep = "_")
}
\arguments{
\item{distribution}{A distribution, inheriting the class
\code{"dst"}, \code{"bidst"}, or \code{"multidst"}.}

\item{at}{A vector of values to evaluate the univariate distributional
representation at.}
}
\value{
The evaluated survival function
in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a distribution's survival function.
}
\examples{
d <- dst_unif(0, 4)
eval_survival(d, at = 0:4)
enframe_survival(d, at = 0:4)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_chf}()},
\code{\link{eval_density}()},
\code{\link{eval_hazard}()},
\code{\link{eval_odds}()},
\code{\link{eval_pmf}()},
\code{\link{eval_quantile}()},
\code{\link{eval_return}()}
}
\concept{distributional representations}
