% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtype.R
\name{vtype}
\alias{vtype}
\title{Variable Type of a Distribution}
\usage{
vtype(distribution)
}
\arguments{
\item{distribution}{Distribution object.}
}
\value{
Single character with the variable type.
}
\description{
Retrieve the variable type of a distribution, such as
"continuous" or "discrete".
}
\examples{
vtype(dst_beta(1, 2))
vtype(dst_bern(0.4))
vtype(distribution(
  cdf = pnorm,
  density = dnorm,
  .support = continuous()
))
}
